+++ src/Attributes/Collections/Values.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\Collections;
+
+use Attribute;
+
+use function array_values;
+
+/**
+ * @psalm-immutable
+ * @author Brian Faust <brian@cline.sh>
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Values extends AbstractCollectionTransformer
+{
+    public function transform(array $items): array
+    {
+        return array_values($items);
+    }
+}